Add instructions for creating bidirectional links
Most apps will automatically create a link to a note that exists within your library.  The image provided seems to suggest this as it contains a link that has the note's URL path.  However, this is not the case and the full link to a note must be used.

This commit adds the additional instruction on how to create the link,  as well as a note about using the entire link.